feat: style profile in small size
