+++ src/pages/ProfilePage/ProfilePage.tsx
@@ -9,24 +9,24 @@ import ProfileChangePassword from "../../components/ProfileChangePassword/Profil
 import Modal from "../../components/Modal/Modal";
 import useModal from "../../hooks/useModal";
 import ModalContentInProfile from "../../components/ModalContentInProfile/ModalContentInProfile";
-import logOut from "../../API/logOut";
 import DeleteAccountImg from "../../assets/png/delete_account.png";
-import LogOutImg from "../../assets/png/log_out.png";
 import deleteAccount from "../../API/deleteAccount";
 import { SnackbarTextValue } from "../../features/SnackbarTextSlice";
 import { ModalContentContext } from "../../contexts/ModalContentContext";
+import LogOutBtn from "./Btns/LogOutBtn";
+import useResize from "../../hooks/useResize";
 
 type ProfileStates = "general" | "changePassword";
 
 export default function ProfilePage() {
   const isAuth = useAppSelector((state) => state.present.authData.value.isAuth);
-  const name = useAppSelector((state) => state.present.authData.value.name);
   const token = useAppSelector((state) => state.present.authData.value.token);
   const isOpen = useAppSelector((state) => state.present.IsOpenModal.value);
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const [profileState, SetProfileState] = useState<ProfileStates>("general");
   const { toggle } = useModal();
+  const { isScreenSm } = useResize();
 
   useEffect(() => {
     if (isAuth !== true) {
@@ -49,7 +49,6 @@ export default function ProfilePage() {
             <article className="profile-page">
               <div className="profile-page__wrapper">
                 <h3 className="profile-page__h3">Profile</h3>
-                {/* TODO: в две строчки */}
                 <section className="profile-menu">
                   <div className="profile-menu__wrapper">
                     <div className="profile-menu__btns">
@@ -73,6 +72,10 @@ export default function ProfilePage() {
                       >
                         Change password
                       </button>
+                      {isScreenSm ? "" : <LogOutBtn />}
+                    </div>
+                    <div className="profile-menu__delete-btn-wrapper">
+                      <LogOutBtn />
                       <button
                         type="button"
                         className="profile-menu__btn"
@@ -80,11 +83,19 @@ export default function ProfilePage() {
                           toggle();
                           setModalContent(
                             <ModalContentInProfile
-                              imageSrc={LogOutImg}
-                              text="Are you sure you want to log out?"
+                              imageSrc={DeleteAccountImg}
+                              text="Are you sure you want to delete your account?"
                               handleClickByOkBtn={() => {
-                                navigate("/");
-                                logOut(name);
+                                deleteAccount(token).then(() => {
+                                  toggle();
+                                  navigate("/");
+                                  dispatch(
+                                    SnackbarTextValue({
+                                      text: "Your account was deleted",
+                                      withUndo: false,
+                                    })
+                                  );
+                                });
                               }}
                               submitBtn={{
                                 text: "Ok",
@@ -95,41 +106,9 @@ export default function ProfilePage() {
                           );
                         }}
                       >
-                        Log out
+                        Delete account
                       </button>
                     </div>
-                    <button
-                      type="button"
-                      className="profile-menu__btn"
-                      onClick={() => {
-                        toggle();
-                        setModalContent(
-                          <ModalContentInProfile
-                            imageSrc={DeleteAccountImg}
-                            text="Are you sure you want to delete your account?"
-                            handleClickByOkBtn={() => {
-                              deleteAccount(token).then(() => {
-                                toggle();
-                                navigate("/");
-                                dispatch(
-                                  SnackbarTextValue({
-                                    text: "Your account was deleted",
-                                    withUndo: false,
-                                  })
-                                );
-                              });
-                            }}
-                            submitBtn={{
-                              text: "Ok",
-                              style: "btn",
-                            }}
-                            cancelBtnStyle="borderNone"
-                          />
-                        );
-                      }}
-                    >
-                      Delete account
-                    </button>
                   </div>
                 </section>
                 <section className="profile-data">
